Initial move to meson build
While makefiles are simple, this specfic setup makes it painful to cross
compile within a buildroot. Using meson should make this "just work".

The host (or generated system) will now require libgpiod, libbluetooth,
and libImlib2 as these were previously provided as binary .so files.

There was a weirdness in one header exporting an inline function in such
a way that the linker just couldn't handle it. Marked that function as
static within the header, and well that worked.

Linking in the logo.png file requied a bit of python magic to make sure
ld was called from the working directory of logo.png rather than
whereever ninja was run from. Hopefully this will work within a
buildroot build.